+++ CarParkingBooking_UI/src/app/Dashboard/Log-Reg/reg/reg.component.html
@@ -71,8 +71,10 @@
             *ngIf="this.validate.validation(this.regpage, 'confirmpassword')"
             >*Confirm password</span
           >
-          <button (click)="signin()">Sign Up</button>
+          <button (click)="signin()">Sign Up
 
+          </button>
+          
           <span class="Dlog"><a href="Dreg">Dealer Register</a></span>
         </form>
       </div>
