+++ CarParkingBooking_UI/src/app/Dashboard/User/e-receipt/e-receipt.component.html
@@ -1,6 +1,6 @@
 
     <body>
-
+<app-razorpaybutton></app-razorpaybutton>
         <div class = "invoice-wrapper" id = "print-area">
             <div class = "invoice">
                 <div class = "invoice-container">
