+++ CarParkingBooking_UI/src/app/Dashboard/User/e-receipt/razorpaybutton/razorpaybutton.component.css
@@ -1,0 +1,18 @@
+.payment-container {
+    text-align: center;
+    margin-top: 50px;
+  }
+  
+  button {
+    padding: 10px 20px;
+    font-size: 16px;
+    background-color: #007bff;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+  }
+  
+  button:hover {
+    background-color: #0056b3;
+  }
