+++ CarParkingBooking_UI/src/app/Dashboard/User/e-receipt/razorpaybutton/razorpaybutton.component.html
@@ -1,0 +1,5 @@
+<div class="payment-container">
+  <h2>UPI Payment</h2>
+  <p>Amount: {{ paymentAmount }} INR</p>
+  <button (click)="initiatePayment()">Pay via UPI</button>
+</div>
