+++ CarParkingBooking_UI/src/app/Dashboard/User/e-receipt/razorpaybutton/razorpaybutton.component.spec.ts
@@ -1,0 +1,28 @@
+/* tslint:disable:no-unused-variable */
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { DebugElement } from '@angular/core';
+
+import { RazorpaybuttonComponent } from './razorpaybutton.component';
+
+describe('RazorpaybuttonComponent', () => {
+  let component: RazorpaybuttonComponent;
+  let fixture: ComponentFixture<RazorpaybuttonComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ RazorpaybuttonComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(RazorpaybuttonComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
