+++ CarParkingBooking_UI/src/app/Dashboard/User/userdata/userdata.component.css
@@ -188,21 +188,3 @@ top:30px;
 width: 200px;
 }
 
-.payment-container {
-  text-align: center;
-  margin-top: 50px;
-}
-
-button {
-  padding: 10px 20px;
-  font-size: 16px;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-button:hover {
-  background-color: #0056b3;
-}
