+++ CarParkingBooking_UI/src/app/Dashboard/User/userdata/userdata.component.html
@@ -15,11 +15,7 @@
       >
         Confirm Booking
       </button>
-      <div class="payment-container">
-        <h2>UPI Payment</h2>
-        <p>Amount: {{ paymentAmount }} INR</p>
-        <button (click)="initiatePayment()">Pay via UPI</button>
-      </div>
+      
     </div>
   </div>
   <div class="right-container">
