+++ CarParkingBooking_UI/src/app/Dashboard/User/userdata/userdata.component.ts
@@ -3,7 +3,7 @@ import { routes } from '../../../app.routes';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { VehicleDetialsService } from '../../../Service/Backend/vehicle-detials.service';
-declare var Razorpay: any;
+
 
 @Component({
   selector: 'app-userdata',
@@ -26,46 +26,6 @@ constructor(
   private backVehicle:VehicleDetialsService) {
   
 }
-
-
-paymentAmount: number = 100; // Amount in INR
-  currency: string = 'INR';
-  upiId: string = 'razorpay.me/@carparking1144';
-  key:string = 'rzp_test_K5F8atqTrPzCOi';
-  initiatePayment() {
-    const options = {
-      key: this.key, // Replace with your Razorpay Key ID
-      amount: this.paymentAmount * 100, // Amount in paise (e.g., 100 INR = 10000 paise)
-      currency: this.currency,
-      name: 'Your Company Name',
-      description: 'Payment for Order',
-      image: 'https://your-company-logo-url.com/logo.png', // Replace with your logo URL
-      handler: (response: any) => {
-        console.log('Payment Successful:', response);
-        alert('Payment Successful! Payment ID: ' + response.razorpay_payment_id);
-        // You can send the payment response to your backend for verification
-      },
-      prefill: {
-        name: 'Customer Name',
-        email: 'customer@example.com',
-        contact: '9999999999',
-        method: 'upi', // Specify UPI as the payment method
-        upi: {
-          vpa: this.upiId // Specify the UPI ID
-        }
-      },
-      notes: {
-        address: 'Customer Address'
-      },
-      theme: {
-        color: '#F37254'
-      }
-    };
-
-    const rzp = new Razorpay(options);
-    rzp.open();
-  }
-
 
   // vechilemodal(){
   //   document.getElementById("myDropdown")!.classList.toggle("show");
